requirements.txt
torch
torchvision
numpy
scikit-learn
joblib
tqdm
pillow
cython
